add mini events and refactored for testing
