+++ .gitignore
@@ -1,1 +1,4 @@
-.vscode/*
+.vscode/
+test/
+node_modules/
+.babelrc
